massively updated README, restructured factions
