+++ README.md
@@ -1,20 +1,29 @@
-**WIP**
-
-## Prerequisites
-
-Highly recommended: `>= npm-3.x`, which comes with NodeJS 6.x. Don't use older versions!
-
-Alternatively, you can install `grad-factions` manually. You will then also need `grad-loadout`.
+_Predefined sets of Arma3 unit loadouts, easily applied._
 
 ## Installation
 
+### Variant 1: npm
+
+Highly recommended: `>= npm-3.x`, which comes with NodeJS 6.x. Don't use older versions!
+
 `npm install grad-factions --save`
 
-## Update
+For updating, change the desired version number in your mission's `package.json`, and do `npm install`
 
-change the desired version number in your mission's `package.json`, and do `npm install`
+### Variant 2: manually
+
+Alternatively, you can install `grad-factions` manually. You will then also need `grad-loadout`.
+
+* download both  `grad-factions` and `grad-loadout` as zipfiles from github, and put the directories into a `node_modules` directories, so that you get:
+
+```
+myMission.altis/node_modules/grad-factions
+myMission.altis/node_modules/grad-loadout
+```
 
 ## Configuration
+
+### description.ext
 
 You *may* set some of those options. They may or may have no effect, depending on the chosen factions.
 
@@ -27,15 +36,16 @@ You *may* set some of those options. They may or may have no effect, depending o
 ```
 Below, add this:
 ```
-#define MODULES_DIRECTORY node_modules
-#include "MODULES_DIRECTORY\grad-factions\loadouts.hpp"
+#include "node_modules\grad-factions\loadouts.hpp"
 
 class CfgFunctions {
-    #include "MODULES_DIRECTORY\grad-loadout\CfgFunctions.hpp"
+    #include "node_modules\grad-loadout\CfgFunctions.hpp"
 };
 ```
 
-You're nearly done. One last thing you will have to do: define which set of factions should actually be used.
+### init.sqf
+
+You're nearly done. One last thing you will have to do: define which loadouts should actually be applied:
 Somewhere in your init.sqf, for example:
 
 ```
@@ -43,19 +53,23 @@ Somewhere in your init.sqf, for example:
 ["OPF_G_F", "ModerateRebels"] call GRAD_Loadout_fnc_FactionSetLoadout;
 ```
 
-These factions are available:
+These factions are available (see also the `factions` directory):
 
-*   "BWFleck"         BLUFOR        Bundeswehr Flecktarn
-*   "BWTrop"          BLUFOR        Bundeswehr Tropentarn
-*   "USOCP"           BLUFOR        US Army OCP
-*   "USUCP"           BLUFOR        US Army UCP
-*   "USMPT"           BLUFOR        US Marines MARPAT
-*   "RUFlora"         OPFOR         Russian MSV Flora
-*   "RUEMR"           OPFOR         Russian MSV EMR-Summer
-*   "ModerateRebels"  INDEPENDENT   Mittelmäßig bewaffnete, halbzivile Rebellen
+| name           |  description                                 |
+|----------------|----------------------------------------------|
+| BwFleck        |  Bundeswehr Flecktarn                        |
+| BwTrop         |  Bundeswehr Tropentarn                       |
+| UsOCP          |  US Army OCP                                 |
+| UsUCP          |  US Army UCP                                 |
+| UsMPT          |  US Marines MARPAT                           |
+| RuFlora        |  Russian MSV Flora                           |
+| RuEMR          |  Russian MSV EMR-Summer                      |
+| ModerateRebels |  Mittelmäßig bewaffnete, halbzivile Rebellen |
 
-**NOTE**: loadouts are defined for vanilla NATO/CSAT/AAF units only – they may or may not work for other units.
-As units differ mostly in their loadouts (which is exchanged completely anyway) this should not matter much.
+
+**HINT**: If you do not know which faction your editor-created units belong to, have a look into the [BI Community Wiki](https://community.bistudio.com/wiki/Arma_3_CfgVehicles_WEST)
+
+**NOTE**: Loadouts are not guaranteed to work with units provided by Mods. If you need a specific mod supported, open an issue with [grad-loadout](https://github.com/gruppe-adler/grad-loadout).
 
 /*  Von den angegebenen Kategorien haben die folgenden Einheiten ein Standardloadout:
 *   Englisch:                                 Deutsch:                                          Funktion:
@@ -78,5 +92,6 @@ As units differ mostly in their loadouts (which is exchanged completely anyway) 
 *   Team Leader                               Teamleader                                        Gruppenführer und Zugführer
 */
 
+--------------------------------------------------------------------------------
 
 http://www.gruppe-adler.de
