+++ factions/ModerateRebels.hpp
