+++ factions/UsUCP.hpp
@@ -1,0 +1,77 @@
+
+#ifdef SUPPRESSORS
+  #define SUPPRESSORITEM "rhsusf_acc_nt4_tan"
+#else
+  #define SUPPRESSORITEM ""
+#endif
+
+class USUCP: USOCP {
+    class AllUnits {
+        uniform = "rhs_uniform_cu_ucp";
+        vest = "rhsusf_iotv_ocp_Rifleman";
+        backpack = "";
+        headgear = "rhsusf_ach_helmet_ucp";
+        primaryWeapon = "rhs_weap_m4a1_blockII_d";
+        primaryWeaponOptics = "FHQ_optic_HWS_G33_tan";
+    };
+    class Type: Type {
+        class Soldier_F: Soldier_F {};
+        class soldier_AAR_F: soldier_AAR_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class support_AMG_F: support_AMG_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_AAA_F: soldier_AAA_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_AAT_F: soldier_AAT_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_AR_F: soldier_AR_F {
+            vest = "rhsusf_iotv_ucp_SAW";
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class medic_F: medic_F {
+            vest = "rhsusf_iotv_ucp_Medic";
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_exp_F: soldier_exp_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class Soldier_GL_F: Soldier_GL_F {
+            primaryWeapon = "rhs_weap_m4a1_blockII_M203_d";
+        };
+        class HeavyGunner_F: HeavyGunner_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_M_F: soldier_M_F {
+            primaryWeapon = "rhs_weap_sr25_ec_d";
+        };
+        class soldier_AA_F: soldier_AA_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_AT_F: soldier_AT_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_repair_F: soldier_repair_F {
+            backpack = "rhsusf_assault_eagleaiii_ucp";
+        };
+        class soldier_LAT_F: soldier_LAT_F {};
+        class Soldier_SL_F: Soldier_SL_F {
+            backpack = "tf_rt1523g_bwmod";
+            headgear = "rhsusf_ach_helmet_headset_ucp";
+            vest = "rhsusf_iotv_ucp_Squadleader";
+        };
+        class Soldier_TL_F: Soldier_TL_F {
+            primaryWeapon = "rhs_weap_m4a1_blockII_M203_d";
+            headgear = "rhsusf_ach_helmet_headset_ucp";
+            vest = "rhsusf_iotv_ucp_Teamleader";
+        };
+    };
+    class Rank: Rank {
+        class LIEUTENANT: LIEUTENANT {
+            headgear = "rhsusf_patrolcap_ucp";
+        };
+    };
+};
